requirements.txt
python-dotenv
openai
mem0ai
fastapi 
uvicorn
sqlalchemy
pydantic
alembic
pytest 
httpx
pytest-asyncio
gunicorn
